feat(web): split Applications tab into Forms and Templates
Replace the single Applications sidebar entry with two focused tabs
under an Applications category:
- Forms: form config, questions, and submissions
- Templates: reusable question sets (built-in + custom)

Remove the now-unused ApplicationsTab.vue.
