+++ web/frontend/src/views/guild/GuildDetailView.vue
@@ -4,7 +4,8 @@ import { useRoute, useRouter } from "vue-router";
 import { Icon } from "@iconify/vue";
 import { useAuthStore } from "@/stores/auth";
 import { useGuildStore } from "@/stores/guild";
-import ApplicationsTab from "./tabs/ApplicationsTab.vue";
+import ApplicationFormsTab from "./tabs/ApplicationFormsTab.vue";
+import ApplicationTemplatesTab from "./tabs/ApplicationTemplatesTab.vue";
 import LanguageTab from "./tabs/LanguageTab.vue";
 import ModeratorRolesTab from "./tabs/ModeratorRolesTab.vue";
 import LeaveMessagesTab from "./tabs/LeaveMessagesTab.vue";
@@ -62,9 +63,10 @@ const sectionGroups = [
     ],
   },
   {
-    label: "Community",
+    label: "Applications",
     items: [
-      { id: "applications", label: "Applications", icon: "mdi:file-document-outline", component: ApplicationsTab },
+      { id: "application-forms", label: "Forms", icon: "mdi:file-document-outline", component: ApplicationFormsTab },
+      { id: "application-templates", label: "Templates", icon: "mdi:file-document-multiple-outline", component: ApplicationTemplatesTab },
     ],
   },
   {
