+++ web/frontend/src/views/guild/tabs/ApplicationFormsTab.vue
@@ -2,11 +2,7 @@
 import { ref, onMounted } from "vue";
 import { Icon } from "@iconify/vue";
 import { api } from "@/api/client";
-import type {
-  ApplicationFormSchema,
-  ApplicationSubmissionSchema,
-  ApplicationTemplateSchema,
-} from "@/api/types";
+import type { ApplicationFormSchema, ApplicationSubmissionSchema } from "@/api/types";
 import DiscordPicker from "@/components/DiscordPicker.vue";
 import { useGuildEntities } from "@/composables/useGuildEntities";
 
@@ -14,9 +10,7 @@ const props = defineProps<{ guildId: string }>();
 
 const { fetchChannels, channelName } = useGuildEntities(props.guildId);
 
-// ── State ──
 const forms = ref<ApplicationFormSchema[]>([]);
-const templates = ref<ApplicationTemplateSchema[]>([]);
 const loading = ref(true);
 const error = ref<string | null>(null);
 
@@ -32,11 +26,6 @@ const editingQuestionId = ref<number | null>(null);
 const editQuestionText = ref("");
 const newQuestionText = ref<Record<number, string>>({});
 
-const showCreateTemplate = ref(false);
-const editingTemplateId = ref<number | null>(null);
-const expandedTemplateId = ref<number | null>(null);
-const newTemplateQuestionText = ref<Record<number, string>>({});
-
 const saving = ref(false);
 const opError = ref<string | null>(null);
 const questionSavedFormId = ref<number | null>(null);
@@ -46,7 +35,6 @@ function flashQuestionSaved(formId: number) {
   setTimeout(() => { questionSavedFormId.value = null; }, 1800);
 }
 
-// Blank form state
 function blankForm() {
   return {
     name: "",
@@ -60,19 +48,14 @@ function blankForm() {
   };
 }
 const formDraft = ref(blankForm());
-const templateDraft = ref({ name: "", approval_message: "", denial_message: "", question_texts: [""] });
 
-// ── Load ──
 onMounted(() => { void load(); void fetchChannels(); });
 
 async function load() {
   loading.value = true;
   error.value = null;
   try {
-    [forms.value, templates.value] = await Promise.all([
-      api.get<ApplicationFormSchema[]>(`/guilds/${props.guildId}/application-forms`),
-      api.get<ApplicationTemplateSchema[]>(`/guilds/${props.guildId}/application-templates`),
-    ]);
+    forms.value = await api.get<ApplicationFormSchema[]>(`/guilds/${props.guildId}/application-forms`);
   } catch (e: unknown) {
     error.value = e instanceof Error ? e.message : "Failed to load";
   } finally {
@@ -80,7 +63,6 @@ async function load() {
   }
 }
 
-// ── Forms ──
 function startEdit(form: ApplicationFormSchema) {
   editingFormId.value = form.id;
   formDraft.value = {
@@ -146,7 +128,6 @@ async function deleteForm(formId: number) {
   }
 }
 
-// ── Questions ──
 async function addQuestion(formId: number) {
   const text = (newQuestionText.value[formId] ?? "").trim();
   if (!text) return;
@@ -191,7 +172,6 @@ async function deleteQuestion(formId: number, questionId: number) {
   }
 }
 
-// ── Submissions ──
 async function loadSubmissions(formId: number) {
   if (expandedSubmissionsFormId.value === formId) {
     expandedSubmissionsFormId.value = null;
@@ -214,85 +194,14 @@ const statusColor: Record<string, string> = {
   approved: "text-green-400",
   denied: "text-destructive",
 };
-
-// ── Templates ──
-
-async function saveTemplate() {
-  saving.value = true;
-  opError.value = null;
-  try {
-    if (editingTemplateId.value !== null) {
-      const updated = await api.put<ApplicationTemplateSchema>(
-        `/guilds/${props.guildId}/application-templates/${editingTemplateId.value}`,
-        { name: templateDraft.value.name, approval_message: templateDraft.value.approval_message || null, denial_message: templateDraft.value.denial_message || null },
-      );
-      const idx = templates.value.findIndex((t) => t.id === editingTemplateId.value);
-      if (idx !== -1) templates.value[idx] = updated;
-      editingTemplateId.value = null;
-    } else {
-      const created = await api.post<ApplicationTemplateSchema>(
-        `/guilds/${props.guildId}/application-templates`,
-        {
-          name: templateDraft.value.name,
-          approval_message: templateDraft.value.approval_message || null,
-          denial_message: templateDraft.value.denial_message || null,
-          question_texts: templateDraft.value.question_texts.filter((q) => q.trim()),
-        },
-      );
-      templates.value.push(created);
-      showCreateTemplate.value = false;
-    }
-  } catch (e: unknown) {
-    opError.value = e instanceof Error ? e.message : "Save failed";
-  } finally {
-    saving.value = false;
-  }
-}
-
-async function deleteTemplate(templateId: number) {
-  if (!confirm("Delete this template?")) return;
-  opError.value = null;
-  try {
-    await api.delete(`/guilds/${props.guildId}/application-templates/${templateId}`);
-    templates.value = templates.value.filter((t) => t.id !== templateId);
-  } catch (e: unknown) {
-    opError.value = e instanceof Error ? e.message : "Delete failed";
-  }
-}
-
-async function addTemplateQuestion(templateId: number) {
-  const text = (newTemplateQuestionText.value[templateId] ?? "").trim();
-  if (!text) return;
-  opError.value = null;
-  try {
-    const q = await api.post(`/guilds/${props.guildId}/application-templates/${templateId}/questions`, { question_text: text });
-    const tpl = templates.value.find((t) => t.id === templateId);
-    if (tpl) tpl.questions.push(q as never);
-    newTemplateQuestionText.value[templateId] = "";
-  } catch (e: unknown) {
-    opError.value = e instanceof Error ? e.message : "Failed to add question";
-  }
-}
-
-async function deleteTemplateQuestion(templateId: number, questionId: number) {
-  opError.value = null;
-  try {
-    await api.delete(`/guilds/${props.guildId}/application-templates/${templateId}/questions/${questionId}`);
-    const tpl = templates.value.find((t) => t.id === templateId);
-    if (tpl) tpl.questions = tpl.questions.filter((q) => q.id !== questionId);
-  } catch (e: unknown) {
-    opError.value = e instanceof Error ? e.message : "Failed to delete question";
-  }
-}
 </script>
 
 <template>
   <div class="space-y-8">
-    <!-- Header -->
     <div class="flex items-center justify-between">
       <div>
-        <h2 class="text-lg font-semibold">Application Forms</h2>
-        <p class="text-muted-foreground text-sm">Manage application forms, questions, and templates.</p>
+        <h2 class="text-lg font-semibold">Forms</h2>
+        <p class="text-muted-foreground text-sm">Manage application forms, questions, and submissions.</p>
       </div>
       <button
         class="bg-primary hover:bg-primary/90 text-primary-foreground px-3 py-1.5 rounded text-sm font-medium transition-colors flex items-center gap-1.5"
@@ -359,7 +268,6 @@ async function deleteTemplateQuestion(templateId: number, questionId: number) {
       <p v-if="forms.length === 0" class="text-muted-foreground text-sm">No forms yet.</p>
 
       <div v-for="form in forms" :key="form.id" class="bg-card border border-border rounded overflow-hidden">
-        <!-- Form header row -->
         <div class="flex items-center gap-3 px-4 py-3">
           <button class="flex-1 flex items-center gap-3 text-left" @click="expandedFormId = expandedFormId === form.id ? null : form.id">
             <Icon
@@ -373,9 +281,7 @@ async function deleteTemplateQuestion(templateId: number, questionId: number) {
           <button class="text-xs text-destructive hover:text-destructive/80 transition-colors" @click="deleteForm(form.id)">Delete</button>
         </div>
 
-        <!-- Expanded content -->
         <div v-if="expandedFormId === form.id" class="border-t border-border px-4 py-4 space-y-5">
-
           <!-- Inline edit panel -->
           <div v-if="editingFormId === form.id" class="space-y-3 pb-3 border-b border-border">
             <p class="text-xs font-semibold text-muted-foreground uppercase tracking-wide">Edit Settings</p>
@@ -435,10 +341,7 @@ async function deleteTemplateQuestion(templateId: number, questionId: number) {
           <div class="space-y-2">
             <div class="flex items-center gap-2">
               <p class="text-xs font-semibold text-muted-foreground uppercase tracking-wide">Questions</p>
-              <span
-                v-if="questionSavedFormId === form.id"
-                class="text-xs text-green-400 transition-opacity"
-              >✓ Saved</span>
+              <span v-if="questionSavedFormId === form.id" class="text-xs text-green-400 transition-opacity">✓ Saved</span>
             </div>
             <p v-if="form.questions.length === 0" class="text-muted-foreground text-xs">No questions yet.</p>
             <div
@@ -469,7 +372,6 @@ async function deleteTemplateQuestion(templateId: number, questionId: number) {
                 >✕</button>
               </template>
             </div>
-            <!-- Add question -->
             <div class="flex gap-2 mt-2">
               <input
                 v-model="newQuestionText[form.id]"
@@ -516,128 +418,6 @@ async function deleteTemplateQuestion(templateId: number, questionId: number) {
                   </div>
                 </div>
               </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- ── Templates ── -->
-    <div class="space-y-4 pt-4 border-t border-border">
-      <div class="flex items-center justify-between">
-        <div>
-          <h3 class="text-base font-semibold">Templates</h3>
-          <p class="text-muted-foreground text-xs">Reusable question sets. Built-in templates are read-only.</p>
-        </div>
-        <button
-          class="bg-muted hover:bg-muted/80 px-3 py-1.5 rounded text-sm transition-colors flex items-center gap-1.5"
-          @click="showCreateTemplate = true; templateDraft = { name: '', approval_message: '', denial_message: '', question_texts: [''] }"
-        >
-          <Icon icon="mdi:plus" class="w-4 h-4" />
-          New Template
-        </button>
-      </div>
-
-      <!-- Create Template Panel -->
-      <div v-if="showCreateTemplate" class="bg-card border border-primary rounded p-4 space-y-3">
-        <p class="text-sm font-semibold">New Template</p>
-        <div class="flex flex-col gap-1">
-          <label class="text-xs text-muted-foreground">Name</label>
-          <input v-model="templateDraft.name" class="bg-input border border-border rounded px-3 py-2 text-sm w-64" />
-        </div>
-        <div class="space-y-1">
-          <label class="text-xs text-muted-foreground">Questions</label>
-          <div v-for="(_, i) in templateDraft.question_texts" :key="i" class="flex gap-2">
-            <input
-              v-model="templateDraft.question_texts[i]"
-              :placeholder="`Question ${i + 1}…`"
-              class="bg-input border border-border rounded px-3 py-1.5 text-sm flex-1"
-            />
-            <button class="text-destructive text-sm" @click="templateDraft.question_texts.splice(i, 1)">✕</button>
-          </div>
-          <button class="text-xs text-primary hover:text-primary/80 transition-colors" @click="templateDraft.question_texts.push('')">+ Add question</button>
-        </div>
-        <div class="flex gap-2">
-          <button
-            class="bg-primary hover:bg-primary/90 text-primary-foreground px-3 py-1.5 rounded text-sm font-medium disabled:opacity-50 transition-colors"
-            :disabled="saving || !templateDraft.name.trim()"
-            @click="saveTemplate"
-          >{{ saving ? "Saving…" : "Create" }}</button>
-          <button class="text-muted-foreground hover:text-foreground text-sm transition-colors" @click="showCreateTemplate = false">Cancel</button>
-        </div>
-      </div>
-
-      <!-- Template list -->
-      <div v-if="!loading" class="space-y-2">
-        <p v-if="templates.length === 0" class="text-muted-foreground text-xs">No templates.</p>
-
-        <div
-          v-for="tpl in templates"
-          :key="tpl.id"
-          class="bg-card border border-border rounded overflow-hidden"
-          :class="{ 'opacity-60': tpl.is_built_in }"
-        >
-          <div class="flex items-center gap-3 px-4 py-2.5">
-            <button class="flex-1 flex items-center gap-2 text-left" @click="expandedTemplateId = expandedTemplateId === tpl.id ? null : tpl.id">
-              <Icon :icon="expandedTemplateId === tpl.id ? 'mdi:chevron-down' : 'mdi:chevron-right'" class="w-4 h-4 text-muted-foreground flex-shrink-0" />
-              <span class="text-sm font-medium">{{ tpl.name }}</span>
-              <span class="text-xs text-muted-foreground">{{ tpl.questions.length }} questions</span>
-              <span v-if="tpl.is_built_in" class="text-xs text-muted-foreground italic">built-in</span>
-            </button>
-            <template v-if="!tpl.is_built_in">
-              <button class="text-xs text-muted-foreground hover:text-foreground transition-colors"
-                @click="editingTemplateId = tpl.id; templateDraft = { name: tpl.name, approval_message: tpl.approval_message ?? '', denial_message: tpl.denial_message ?? '', question_texts: [] }; expandedTemplateId = tpl.id"
-              >Edit</button>
-              <button class="text-xs text-destructive hover:text-destructive/80 transition-colors" @click="deleteTemplate(tpl.id)">Delete</button>
-            </template>
-          </div>
-
-          <div v-if="expandedTemplateId === tpl.id" class="border-t border-border px-4 py-3 space-y-3">
-            <!-- Inline edit for guild template -->
-            <div v-if="editingTemplateId === tpl.id && !tpl.is_built_in" class="space-y-2 pb-3 border-b border-border">
-              <div class="flex flex-col gap-1">
-                <label class="text-xs text-muted-foreground">Name</label>
-                <input v-model="templateDraft.name" class="bg-input border border-border rounded px-3 py-2 text-sm w-64" />
-              </div>
-              <div class="flex gap-2">
-                <button
-                  class="bg-primary hover:bg-primary/90 text-primary-foreground px-3 py-1.5 rounded text-sm font-medium disabled:opacity-50 transition-colors"
-                  :disabled="saving"
-                  @click="saveTemplate"
-                >{{ saving ? "Saving…" : "Save" }}</button>
-                <button class="text-muted-foreground hover:text-foreground text-sm transition-colors" @click="editingTemplateId = null">Cancel</button>
-              </div>
-            </div>
-
-            <!-- Questions list -->
-            <ol class="space-y-1">
-              <li
-                v-for="q in [...tpl.questions].sort((a, b) => a.sort_order - b.sort_order)"
-                :key="q.id"
-                class="flex items-start gap-2 group text-sm"
-              >
-                <span class="text-muted-foreground text-xs w-5 flex-shrink-0 pt-0.5">{{ q.sort_order }}.</span>
-                <span class="flex-1">{{ q.question_text }}</span>
-                <button
-                  v-if="!tpl.is_built_in"
-                  class="text-xs text-destructive hover:text-destructive/80 opacity-0 group-hover:opacity-100 transition-opacity flex-shrink-0"
-                  @click="deleteTemplateQuestion(tpl.id, q.id)"
-                >✕</button>
-              </li>
-            </ol>
-            <!-- Add question (guild templates only) -->
-            <div v-if="!tpl.is_built_in" class="flex gap-2">
-              <input
-                v-model="newTemplateQuestionText[tpl.id]"
-                placeholder="New question…"
-                class="bg-input border border-border rounded px-3 py-1.5 text-sm flex-1"
-                @keyup.enter="addTemplateQuestion(tpl.id)"
-              />
-              <button
-                class="bg-muted hover:bg-muted/80 px-3 py-1.5 rounded text-sm transition-colors disabled:opacity-50"
-                :disabled="!(newTemplateQuestionText[tpl.id] ?? '').trim()"
-                @click="addTemplateQuestion(tpl.id)"
-              >Add</button>
             </div>
           </div>
         </div>
